chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-host_sources-ports-expected.txt

This is a testharness.js-based test.
FAIL Specified ports must match. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Returned CSP should be subsumed even if the port is not specified but is a default port for a scheme.
PASS Returned CSP should be subsumed even if the port is not specified but is a default port for a more secure scheme.
PASS The same should hold for `ws` case.
PASS Unspecified ports must match if schemes match.
PASS Returned CSP should be subsumed if the port is specified.
PASS Returned CSP should be subsumed if the port is specified but the scheme is more secure.
FAIL Returned CSP should be subsumed if the port is specified but is not default for a more secure scheme. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Returned CSP should be subsumed if the ports match but schemes are not identical.
PASS Returned CSP should be subsumed if the ports match but schemes are not identical for `ws`.
PASS Wildcard port should match unspecified port.
PASS Wildcard port should match any specific port.
PASS Wildcard port should match a wildcard.
FAIL Wildcard port should not be subsumed by a default port. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Wildcard port should not be subsumed by a spcified port. assert_unreached: No message should be sent from the frame. Reached unreachable code
Harness: the test ran to completion.
